Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-generated-content-expected.txt

A
Y A
XYX

FAIL fieldset generated content assert_equals: clientWidth expected 59 but got 47
